Formula builder component refactored into smaller components
